Move generic build-only tests to their own separate file
